Implemented log-before-write. If writing to the QLDB fails, the endpoint returns a 500 error and does not apply any changes requested from the user, if applicable. Most of the changes in the services are procedural ones, but EncryptionKeyService.java required a fair amount of reorganizing to implement. No test cases were changed, so the end behavior should still be the same assuming writing to QLDB is successful.
Implemented itemType and itemKey as top-level fields for the purposes of indexing on the QLDB. Also implemented indexing on the QLDB, which is necessary to keep infra costs manageable at large database sizes.

Added a new QLDB entry for the null item. This entry is written to in the case where a client writes (correctly) to an endpoint expecting possibly multiple items in the S3 to be modified, but no item is actually modified. Currently, only a single service (EncryptionKeyService.java) writes to the item in the QLDB.