Inject auth provider using a new singletons provider.
Tidy up inconsistent comment formats.
Allow the admin verticle to start without a v2 API.
Add auth to the new site ID router.
Switch to passing the parent router in to the v2 router - that way the admin Verticle doesn't need to know the path the v2 router wants to use.